+++ app/src/main/kotlin/io/github/pawgli/daggercomposelibrary/MainActivity.kt
@@ -13,34 +13,34 @@ import androidx.compose.ui.tooling.preview.Preview
 import io.github.pawgli.daggercomposelibrary.ui.theme.DaggerComposeLibraryTheme
 
 class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            DaggerComposeLibraryTheme {
-                // A surface container using the 'background' color from the theme
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colorScheme.background
-                ) {
-                    Greeting("Android")
-                }
-            }
+  override fun onCreate(savedInstanceState: Bundle?) {
+    super.onCreate(savedInstanceState)
+    setContent {
+      DaggerComposeLibraryTheme {
+        // A surface container using the 'background' color from the theme
+        Surface(
+          modifier = Modifier.fillMaxSize(),
+          color = MaterialTheme.colorScheme.background
+        ) {
+          Greeting("Android")
         }
+      }
     }
+  }
 }
 
 @Composable
 fun Greeting(name: String, modifier: Modifier = Modifier) {
-    Text(
-        text = "Hello $name!",
-        modifier = modifier
-    )
+  Text(
+    text = "Hello $name!",
+    modifier = modifier
+  )
 }
 
 @Preview(showBackground = true)
 @Composable
 fun GreetingPreview() {
-    DaggerComposeLibraryTheme {
-        Greeting("Android")
-    }
+  DaggerComposeLibraryTheme {
+    Greeting("Android")
+  }
 }
