+++ app/src/test/java/io/github/pawgli/daggercomposelibrary/ExampleUnitTest.kt
@@ -1,0 +1,13 @@
+package io.github.pawgli.daggercomposelibrary
+
+import io.kotest.matchers.shouldNotBe
+import org.junit.jupiter.api.Test
+
+class ExampleUnitTest {
+
+  @Test
+  fun `example test`() {
+    2 shouldNotBe 3
+  }
+
+}
