+++ build.gradle.kts
@@ -1,4 +1,4 @@
 plugins {
-    alias(libs.plugins.android.app) apply false
-    alias(libs.plugins.kotlin.android) apply false
+  alias(libs.plugins.android.app) apply false
+  alias(libs.plugins.kotlin.android) apply false
 }
