+++ gradle/libs.versions.toml
@@ -14,9 +14,11 @@ compose-uiPreview = { module = "androidx.compose.ui:ui-tooling-preview" }
 material3 = { module = "androidx.compose.material3:material3" }
 
 junit-jupiter = { module = "org.junit.jupiter:junit-jupiter", version = "5.10.0" }
+kotest = { module = "io.kotest:kotest-assertions-core", version = "5.7.2" }
+
+[bundles]
+testSuite = ["junit-jupiter", "kotest"]
 
 [plugins]
 android-app = { id = "com.android.application", version.ref = "androidGradlePlugin" }
 kotlin-android = { id = "org.jetbrains.kotlin.android", version.ref = "kotlin" }
-
-[bundles]
